refactor: rename component AccoutCard to AccountCard
